Refactor dreamer-kmm boundary and models
- Extract duplicate fitness and maxDrawdown extensions into `BacktestMetricsUtils.kt`.
- Decouple `StochasticBagSpanTrainer` from hardcoded synthetic data by moving the generation logic to tests (`SyntheticData.kt`).
- Extract duplicate `TrikeAdapterLocal.kt` components.
- Delete dead `CursorBacktestAdapters.kt` shim.
- Make paper trading models (`PaperAccount`, `PaperOrder`, `PaperPosition`) plain data classes without `AsyncContextElement` bindings.
- Extract `RowVec` extensions (e.g. `doubleValue`) into `RowVecExtensions.kt`.
- Refactor `TradingEngine` state to rely on an immutable `EngineState`.
- Document thread-safety guarantees on `SimWallet` and `TradingEngine`.
