+++ libs/dreamer-kmm/src/commonMain/kotlin/borg/trikeshed/dreamer/BacktestModels.kt
@@ -4,11 +4,16 @@ import borg.trikeshed.cursor.Cursor
 import borg.trikeshed.lib.*
 import borg.trikeshed.miniduck.getValue
 import borg.trikeshed.cursor.RowVec
+import borg.trikeshed.cursor.stringValue
+import borg.trikeshed.cursor.longValue
+import borg.trikeshed.cursor.doubleValue
+import borg.trikeshed.cursor.intValue
 import borg.trikeshed.cursor.at
 import kotlin.math.sqrt
 
 /**
  * Back-test input derived from a single kline bar.
+
  *
  * [PortfolioInput] is what the [TradingEngine] consumes on each cycle:
  * a list of [PortfolioRow] one per symbol, with current price and quantity.
@@ -552,23 +557,3 @@ suspend fun simulateMultiSymbolTicks(
             avgHarvestPerTick = 0.0,
         ),
     )
- public fun RowVec.stringValue(name: String, default: String): String =
-    getValue(name) as? String ?: default
- public fun RowVec.longValue(name: String): Long = when (val value = getValue(name)) {
-    is Long -> value
-    is Number -> value.toLong()
-    is String -> value.toLongOrNull() ?: 0L
-    else -> 0L
-}
- public fun RowVec.doubleValue(name: String): Double = when (val value = getValue(name)) {
-    is Double -> value
-    is Number -> value.toDouble()
-    is String -> value.toDoubleOrNull() ?: 0.0
-    else -> 0.0
-}
- public fun RowVec.intValue(name: String): Int = when (val value = getValue(name)) {
-    is Int -> value
-    is Number -> value.toInt()
-    is String -> value.toIntOrNull() ?: 0
-    else -> 0
-}
