+++ libs/dreamer-kmm/src/commonMain/kotlin/borg/trikeshed/dreamer/ControlHarness.kt
@@ -1,6 +1,13 @@
 package borg.trikeshed.dreamer
 
+import borg.trikeshed.cursor.intValue
+
+import borg.trikeshed.cursor.doubleValue
+
+import borg.trikeshed.cursor.stringValue
+
 import borg.trikeshed.context.AsyncContextElement
+import borg.trikeshed.cursor.longValue
 import borg.trikeshed.context.AsyncContextKey
 import borg.trikeshed.context.ElementState
 import borg.trikeshed.cursor.Cursor
