+++ libs/dreamer-kmm/src/commonMain/kotlin/borg/trikeshed/dreamer/PaperPosition.kt
@@ -1,8 +1,5 @@
 package borg.trikeshed.dreamer
 
-import borg.trikeshed.context.AsyncContextElement
-import borg.trikeshed.context.AsyncContextKey
-
 /**
  * Simulated open position tracking.
  *
@@ -16,10 +13,7 @@ data class PaperPosition(
     val symbol: String,
     val quantity: Double,
     val entryPrice: Double,
-) : AsyncContextElement() {
-    companion object Key : AsyncContextKey<PaperPosition>()
-    override val key: AsyncContextKey<PaperPosition> get() = Key
-
+) {
     /**
      * Unrealized PnL = (currentPrice - entryPrice) * quantity.
      * Positive when price is above entry, negative when below.
