+++ libs/dreamer-kmm/src/commonMain/kotlin/borg/trikeshed/dreamer/StochasticBag.kt
@@ -1,6 +1,10 @@
 package borg.trikeshed.dreamer
 
 import borg.trikeshed.cursor.Cursor
+import borg.trikeshed.cursor.longValue
+import borg.trikeshed.cursor.intValue
+import borg.trikeshed.cursor.doubleValue
+import borg.trikeshed.cursor.stringValue
 import borg.trikeshed.cursor.at
 import borg.trikeshed.lib.size
 import borg.trikeshed.miniduck.columnar.SpanMatcher
